services/ocr-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-cloud-documentai==3.6.0
pydantic==2.7.4
python-multipart==0.0.9
